+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Version counting is based on semantic versioning (Major.Feature.Patch)
 ### YACReader
 * Improve save current page. It includes the comic file name in the default image name and the page number.
 * Add a extract page action. It extracts the original visible pages when image based variants are used. File name and page number is used as the output. 
+* Save the destination path for save/export page between runs.
 * Fix magnifying glass not using the source material.
 * Improve last page detection in continuous scroll mode.
 * Improve What's new dialog to show the full changelog.